+++ R/fw13_downloadData.R
@@ -10,7 +10,7 @@
 #' single character string containing the RAWS output.
 #'
 #' Station identifiers can be found at https://cefa.dri.edu/raws/RAWSfw13list.xlsx
-#' and can be downloaded with \code{\link{fw13_createMetadata}}.
+#' and can be downloaded with \code{\link{fw13_createMeta}}.
 #' 
 #' @return String containing RAWS data.
 #'