+++ R/wrcc_parseData.R
@@ -5,19 +5,22 @@
 #'
 #' @param fileString character string containing RAWS data
 #' @description Raw character data from WRCC are parsed into a tibble.
-#' The incoming \code{fileString}
-#' can be read in directly from WRCC using \code{wrcc_downloadData()} or from a local
-#' file using \code{readr::read_file()}.
+#' The incoming \code{fileString} can be read in directly from WRCC using 
+#' \code{wrcc_downloadData()} or from a local file using \code{readr::read_file()}.
 #'
-#' The type of monitor represented by this fileString is inferred from the column names
-#' using \code{wrcc_identifyMonitorType()} and appropriate column types are assigned.
-#' The character data are then processed, read into a tibble and augmented in the following ways:
+#' The type of monitor represented by this fileString is inferred from the 
+#' column names using \code{wrcc_identifyMonitorType()} and appropriate column 
+#' types are assigned. The character data are then processed, read into a tibble 
+#' and augmented in the following ways:
+#' 
 #' \enumerate{
-#' \item{Spaces at the beginning and end of each line are moved.}
+#' \item{Spaces at the beginning and end of each line are removed.}
 #' \item{All header lines beginning with ':' are removed.}
 #' }
+#' 
 #' @return Dataframe of RAWS raw station data.
 #' @references \href{https://raws.dri.edu/}{RAWS USA Climate Archive}
+#' 
 #' @examples
 #' \dontrun{
 #' library(RAWSmet)
@@ -26,11 +29,19 @@
 #' tbl <- wrcc_parseData(fileString)
 #' }
 
-wrcc_parseData <- function(fileString) {
-
+wrcc_parseData <- function(
+  fileString = NULL
+) {
+  
+  # ----- Validate parameters --------------------------------------------------
+  
+  MazamaCoreUtils::stopIfNull(fileString)
+  
+  # ----- Identify format ------------------------------------------------------
+  
   if( MazamaCoreUtils::logger.isInitialized() )
     logger.debug(" ----- wrcc_parseData() ----- ")
-
+  
   # Identify monitor type
   monitorTypeList <- wrcc_identifyMonitorType(fileString)
   
@@ -43,6 +54,9 @@ wrcc_parseData <- function(fileString) {
   rawNames <- monitorTypeList$rawNames
   columnNames <- monitorTypeList$columnNames
   columnTypes <- monitorTypeList$columnTypes
+  columnUnits <- monitorTypeList$columnUnits
+  
+  # ----- Parse fileString -----------------------------------------------------
   
   # Convert the fileString into individual lines
   lines <- readr::read_lines(fileString)
@@ -79,7 +93,69 @@ wrcc_parseData <- function(fileString) {
     dplyr::na_if(-9999)
   
   tbl$monitorType <- monitorType
-
+  
+  # ----- Convert to metric units ----------------------------------------------
+  
+  # Precip
+  unit <- columnUnits[which(columnNames == "precipitation")]
+  if ( unit != "mm" ) {
+    stop(sprintf("Need to handle precipitation units of \"%s\" in wrcc_parseData()"))
+  }
+  
+  # Temperature
+  unit <- columnUnits[which(columnNames == "temperature")]
+  if ( unit != "degC" ) {
+    stop(sprintf("Need to handle temperature units of \"%s\" in wrcc_parseData()"))
+  }
+  
+  # Fuel Temperature
+  unit <- columnUnits[which(columnNames == "fuelTemperature")]
+  if ( unit != "degC" ) {
+    stop(sprintf("Need to handle fuelTemperature units of \"%s\" in wrcc_parseData()"))
+  }
+  
+  # Wind Speed
+  unit <- columnUnits[which(columnNames == "windSpeed")]
+  if ( unit != "m/s" ) {
+    stop(sprintf("Need to handle windSpeed units of \"%s\" in wrcc_parseData()"))
+  }
+  
+  # Max Gust Speed
+  unit <- columnUnits[which(columnNames == "maxGustSpeed")]
+  if ( unit != "m/s" ) {
+    stop(sprintf("Need to handle maxGustSpeed units of \"%s\" in wrcc_parseData()"))
+  }
+ 
+  # ----- Convert precipitation from "cumulative since water year start" -------
+  
+  tbl$precipitation <- c(NA, diff(tbl$precipitation))
+  # Handle water year reset.
+  tbl$precipitation[tbl$precipitation < 0] <- 0
+  
+  # ----- Return ---------------------------------------------------------------
+  
   return(tbl)
+  
+}
+
+# ===== DEBUGGING ==============================================================
+
+if ( FALSE ) {
+  
+  library(RAWSmet)
+  
+  wrccID <- "OROB"
+  startdate <- 20200101
+  enddate <- 20200201
+  password <- "MY_PASSWORD"
+  
+  
+  fileString <- wrcc_downloadData(
+    wrccID = wrccID,
+    startdate = startdate,
+    enddate = enddate,
+    password = password
+  )
+  
   
 }