+++ local_eli/attempt_wind_map.R
@@ -3,7 +3,7 @@ library(dplyr)
 library(leaflet)
 library(leaflet.minicharts)
 
-fw13Meta <- fw13_createMetadata()
+fw13Meta <- fw13_createMeta()
 
 waStationIDs <- fw13Meta$nwsID[fw13Meta$stateCode == "WA" & !is.na(fw13Meta$stateCode)]
 allDataWA <- lapply(waStationIDs, fw13_createTimeseriesObject, fw13Meta)